separated list and dict services
